[refactor] #77 detail package 수정
